added more npm commands; moved src folder contents into client folder; added axios to client dependencies
