+++ client/src/App.js
@@ -1,13 +1,11 @@
 import React, { Component } from "react";
 import {
   BrowserRouter as Router,
-  Redirect,
   Route,
   Switch
 } from "react-router-dom";
-import Landing from "./pages/Landing";
+import Landing from "./pages/Landing/Landing";
 import About from "./pages/About/About";
-// import NavBar from "./components/NavBar/NavBar";
 import "./styles/App.sass";
 
 class App extends Component {
