+++ client/src/components/Modal/Modal.scss
