+++ client/src/pages/Landing/Landing.jsx
@@ -1,8 +1,8 @@
 import React from "react";
-import NavBar from "../components/NavBar/NavBar";
-import Modal from "../components/Modal/Modal";
-import ModalDemo from "../components/Modal/ModalDemo";
-import ProductGrid from "../components/ProductGrid/ProductGrid";
+import NavBar from "../../components/NavBar/NavBar";
+import Modal from "../../components/Modal/Modal";
+import ModalDemo from "../../components/Modal/ModalDemo";
+import ProductGrid from "../../components/ProductGrid/ProductGrid";
 
 
 const Landing = props => {
