+++ package.json
@@ -1,24 +1,18 @@
 {
   "name": "bears-project-9",
-<<<<<<< HEAD
   "version": "1.0.0",
   "description": "chingu voyage",
-  "main": "index.js",
-=======
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "react": "^16.6.3",
-    "react-dom": "^16.6.3",
-    "react-router-dom": "^4.3.1",
-    "react-scripts": "2.1.1"
-  },
->>>>>>> 853c83549c73df255be3d26ce4feae9c7d0d4c9b
+  "main": "server.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "seed": "node seed.js",
     "resetDb": "node resetDb.js",
-    "dev": "nodemon server.js"
+    "dev": "nodemon server.js",
+    "start": "if-env NODE_ENV=production && npm run start:prod || npm run start:dev",
+    "start:prod": "node server.js",
+    "start:dev": "concurrently \"nodemon --ignore 'client/*'\" \"npm run client\"",
+    "install": "cd client && npm install",
+    "client": "cd client && npm run start"
   },
   "repository": {
     "type": "git",
@@ -34,23 +28,14 @@
     "body-parser": "^1.18.3",
     "cors": "^2.8.5",
     "express": "^4.16.4",
+    "if-env": "^1.0.4",
     "morgan": "^1.9.1",
     "pg": "^7.7.1",
     "pg-hstore": "^2.3.2",
     "sequelize": "^4.42.0"
   },
-<<<<<<< HEAD
   "devDependencies": {
+    "concurrently": "^3.5.0",
     "nodemon": "^1.18.9"
-=======
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ],
-  "devDependencies": {
-    "node-sass": "^4.11.0"
->>>>>>> 853c83549c73df255be3d26ce4feae9c7d0d4c9b
   }
 }
